Renamed search folder in models
