+++ packages/models/src/global-search/index.ts
