+++ packages/models/src/global-search/searchResponse.ts
