+++ packages/models/src/index.ts
@@ -16,4 +16,4 @@ export * from './order/index';
 export * from './product/index';
 export * from './user/index';
 export * from './tags/index';
-export * from './elastic-search/index'
+export * from './global-search/index'
